% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{acl_config_from_list}
\alias{acl_config_from_list}
\title{Create ACL Config from List}
\usage{
acl_config_from_list(acl_list)
}
\arguments{
\item{acl_list}{Character vector of ACL strings}
}
\value{
ACLConfig object
}
\description{
Create ACLConfig from list of ACL strings
}
