% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{ACLConfig}
\alias{ACLConfig}
\title{ACL Configuration}
\description{
ACL configuration for a document or collection
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{read_principals}}{Who can read}

\item{\code{deny_principals}}{Who cannot read (takes precedence)}

\item{\code{is_public}}{Is public access allowed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ACLConfig-new}{\code{ACLConfig$new()}}
\item \href{#method-ACLConfig-clone}{\code{ACLConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ACLConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-ACLConfig-new}{}}}
\subsection{Method \code{new()}}{
Create a new ACLConfig
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ACLConfig$new(
  read_principals = list(),
  deny_principals = list(),
  is_public = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{read_principals}}{List of ACLPrincipal objects}

\item{\code{deny_principals}}{List of ACLPrincipal objects}

\item{\code{is_public}}{Logical}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ACLConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ACLConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ACLConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
