% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\docType{data}
\name{ACLOperator}
\alias{ACLOperator}
\title{ACL Operator Types}
\format{
An object of class \code{list} of length 5.
}
\usage{
ACLOperator
}
\description{
ACL matching operators
}
\keyword{datasets}
