% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{ACLPrincipal}
\alias{ACLPrincipal}
\title{ACL Principal}
\description{
An ACL principal (user, group, or role)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{Principal type}

\item{\code{value}}{Principal value}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ACLPrincipal-new}{\code{ACLPrincipal$new()}}
\item \href{#method-ACLPrincipal-matches}{\code{ACLPrincipal$matches()}}
\item \href{#method-ACLPrincipal-to_string}{\code{ACLPrincipal$to_string()}}
\item \href{#method-ACLPrincipal-clone}{\code{ACLPrincipal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ACLPrincipal-new"></a>}}
\if{latex}{\out{\hypertarget{method-ACLPrincipal-new}{}}}
\subsection{Method \code{new()}}{
Create a new ACLPrincipal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ACLPrincipal$new(type, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Principal type (user, group, role)}

\item{\code{value}}{Principal value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ACLPrincipal-matches"></a>}}
\if{latex}{\out{\hypertarget{method-ACLPrincipal-matches}{}}}
\subsection{Method \code{matches()}}{
Check if this principal matches another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ACLPrincipal$matches(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{Another ACLPrincipal}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ACLPrincipal-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-ACLPrincipal-to_string}{}}}
\subsection{Method \code{to_string()}}{
Convert to string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ACLPrincipal$to_string()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ACLPrincipal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ACLPrincipal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ACLPrincipal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
