% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{advanced_search}
\alias{advanced_search}
\title{VectrixDB Advanced Search Features}
\description{
Enterprise-grade search capabilities:
\itemize{
\item Faceted search with aggregations
\item ACL/Security filtering
\item Text analyzers (stemming, synonyms, stopwords)
}
}
