% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{AnalyzerChain}
\alias{AnalyzerChain}
\title{Analyzer Chain}
\description{
Chain multiple analyzers together
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{analyzers}}{List of TextAnalyzer objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnalyzerChain-new}{\code{AnalyzerChain$new()}}
\item \href{#method-AnalyzerChain-analyze}{\code{AnalyzerChain$analyze()}}
\item \href{#method-AnalyzerChain-clone}{\code{AnalyzerChain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalyzerChain-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnalyzerChain-new}{}}}
\subsection{Method \code{new()}}{
Create a new AnalyzerChain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalyzerChain$new(analyzers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyzers}}{List of TextAnalyzer objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalyzerChain-analyze"></a>}}
\if{latex}{\out{\hypertarget{method-AnalyzerChain-analyze}{}}}
\subsection{Method \code{analyze()}}{
Run text through all analyzers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalyzerChain$analyze(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{Input text}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector of tokens
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalyzerChain-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnalyzerChain-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalyzerChain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
