% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_config_from_env}
\alias{cache_config_from_env}
\title{Create Cache Config from Environment}
\usage{
cache_config_from_env()
}
\value{
CacheConfig object
}
\description{
Create config from environment variables
}
