% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{CacheBackend}
\alias{CacheBackend}
\title{Cache Backend Types}
\format{
An object of class \code{list} of length 3.
}
\usage{
CacheBackend
}
\description{
Available cache backends
}
\keyword{datasets}
