% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{CacheEntry}
\alias{CacheEntry}
\title{Cache Entry}
\description{
A cached entry with metadata
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Cached value}

\item{\code{created_at}}{Creation timestamp}

\item{\code{ttl}}{Time to live in seconds}

\item{\code{hits}}{Number of hits}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CacheEntry-new}{\code{CacheEntry$new()}}
\item \href{#method-CacheEntry-is_expired}{\code{CacheEntry$is_expired()}}
\item \href{#method-CacheEntry-clone}{\code{CacheEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-CacheEntry-new}{}}}
\subsection{Method \code{new()}}{
Create a new CacheEntry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheEntry$new(value, ttl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value to cache}

\item{\code{ttl}}{Time to live in seconds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheEntry-is_expired"></a>}}
\if{latex}{\out{\hypertarget{method-CacheEntry-is_expired}{}}}
\subsection{Method \code{is_expired()}}{
Check if entry is expired
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheEntry$is_expired()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CacheEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
