% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{CacheStats}
\alias{CacheStats}
\title{Cache Statistics}
\description{
Cache statistics for monitoring
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{hits}}{Cache hits}

\item{\code{misses}}{Cache misses}

\item{\code{sets}}{Cache sets}

\item{\code{deletes}}{Cache deletes}

\item{\code{evictions}}{Cache evictions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CacheStats-record_hit}{\code{CacheStats$record_hit()}}
\item \href{#method-CacheStats-record_miss}{\code{CacheStats$record_miss()}}
\item \href{#method-CacheStats-record_set}{\code{CacheStats$record_set()}}
\item \href{#method-CacheStats-record_delete}{\code{CacheStats$record_delete()}}
\item \href{#method-CacheStats-record_eviction}{\code{CacheStats$record_eviction()}}
\item \href{#method-CacheStats-hit_rate}{\code{CacheStats$hit_rate()}}
\item \href{#method-CacheStats-to_list}{\code{CacheStats$to_list()}}
\item \href{#method-CacheStats-reset}{\code{CacheStats$reset()}}
\item \href{#method-CacheStats-clone}{\code{CacheStats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheStats-record_hit"></a>}}
\if{latex}{\out{\hypertarget{method-CacheStats-record_hit}{}}}
\subsection{Method \code{record_hit()}}{
Record a cache hit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheStats$record_hit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheStats-record_miss"></a>}}
\if{latex}{\out{\hypertarget{method-CacheStats-record_miss}{}}}
\subsection{Method \code{record_miss()}}{
Record a cache miss
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheStats$record_miss()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheStats-record_set"></a>}}
\if{latex}{\out{\hypertarget{method-CacheStats-record_set}{}}}
\subsection{Method \code{record_set()}}{
Record a cache set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheStats$record_set()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheStats-record_delete"></a>}}
\if{latex}{\out{\hypertarget{method-CacheStats-record_delete}{}}}
\subsection{Method \code{record_delete()}}{
Record a cache delete
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheStats$record_delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheStats-record_eviction"></a>}}
\if{latex}{\out{\hypertarget{method-CacheStats-record_eviction}{}}}
\subsection{Method \code{record_eviction()}}{
Record a cache eviction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheStats$record_eviction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheStats-hit_rate"></a>}}
\if{latex}{\out{\hypertarget{method-CacheStats-hit_rate}{}}}
\subsection{Method \code{hit_rate()}}{
Get hit rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheStats$hit_rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric hit rate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheStats-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CacheStats-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheStats$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List representation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheStats-reset"></a>}}
\if{latex}{\out{\hypertarget{method-CacheStats-reset}{}}}
\subsection{Method \code{reset()}}{
Reset statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheStats$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheStats-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CacheStats-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
