% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_python_module}
\alias{check_python_module}
\title{Check if Python module is available}
\usage{
check_python_module(module)
}
\arguments{
\item{module}{Module name}
}
\value{
Logical
}
\description{
Check if Python module is available
}
\keyword{internal}
