% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_print}
\alias{cli_print}
\title{Print CLI Message}
\usage{
cli_print(msg, type = "info")
}
\arguments{
\item{msg}{Message}

\item{type}{Type: "info", "success", "warning", "error"}
}
\description{
Print CLI Message
}
\keyword{internal}
