% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{CLIConfig}
\alias{CLIConfig}
\title{CLI Configuration}
\description{
Configuration for CLI behavior
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{verbose}}{Print verbose output}

\item{\code{color}}{Use colored output}

\item{\code{data_dir}}{Default data directory}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CLIConfig-new}{\code{CLIConfig$new()}}
\item \href{#method-CLIConfig-clone}{\code{CLIConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CLIConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-CLIConfig-new}{}}}
\subsection{Method \code{new()}}{
Create CLI config
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CLIConfig$new(verbose = TRUE, color = TRUE, data_dir = "./vectrixdb_data")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Verbose output}

\item{\code{color}}{Colored output}

\item{\code{data_dir}}{Data directory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CLIConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CLIConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CLIConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
