% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\name{Community}
\alias{Community}
\title{Community}
\description{
A community of entities
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Community ID}

\item{\code{level}}{Hierarchy level}

\item{\code{entity_ids}}{Member entity IDs}

\item{\code{summary}}{Community summary}

\item{\code{parent_id}}{Parent community ID}

\item{\code{child_ids}}{Child community IDs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Community-new}{\code{Community$new()}}
\item \href{#method-Community-size}{\code{Community$size()}}
\item \href{#method-Community-clone}{\code{Community$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Community-new"></a>}}
\if{latex}{\out{\hypertarget{method-Community-new}{}}}
\subsection{Method \code{new()}}{
Create a new Community
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Community$new(
  id,
  level = 0,
  entity_ids = character(0),
  summary = NULL,
  parent_id = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID}

\item{\code{level}}{Level}

\item{\code{entity_ids}}{Members}

\item{\code{summary}}{Summary}

\item{\code{parent_id}}{Parent}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Community-size"></a>}}
\if{latex}{\out{\hypertarget{method-Community-size}{}}}
\subsection{Method \code{size()}}{
Get size
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Community$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Community-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Community-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Community$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
