% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Compute cosine similarity}
\usage{
cosine_similarity(query_vector, doc_vectors)
}
\arguments{
\item{query_vector}{Numeric vector (query)}

\item{doc_vectors}{Matrix of document vectors}
}
\value{
Numeric vector of similarities
}
\description{
Compute cosine similarity
}
\keyword{internal}
