% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{create_cache}
\alias{create_cache}
\title{Create Cache}
\usage{
create_cache(config = NULL)
}
\arguments{
\item{config}{CacheConfig object or NULL for defaults}
}
\value{
Cache object
}
\description{
Factory function to create cache backend
}
