% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\name{create_default_graphrag_config}
\alias{create_default_graphrag_config}
\title{Create Default GraphRAG Config}
\usage{
create_default_graphrag_config(...)
}
\arguments{
\item{...}{Additional options}
}
\value{
GraphRAGConfig object
}
\description{
Create default config with regex extractor
}
