% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnsw.R
\name{create_hnsw_index}
\alias{create_hnsw_index}
\title{Create HNSW Index}
\usage{
create_hnsw_index(dimension, metric = "angular", n_trees = 50)
}
\arguments{
\item{dimension}{Vector dimension}

\item{metric}{Distance metric}

\item{n_trees}{Number of trees}
}
\value{
HNSWIndex object
}
\description{
Factory function to create HNSW index
}
