% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnsw.R, R/types.R
\docType{data}
\name{DistanceMetric}
\alias{DistanceMetric}
\title{Distance Metric Types}
\format{
An object of class \code{list} of length 4.

An object of class \code{list} of length 4.
}
\usage{
DistanceMetric

DistanceMetric
}
\description{
Available distance metrics for vector comparison
}
\keyword{datasets}
