% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\name{DocumentChunker}
\alias{DocumentChunker}
\title{Document Chunker}
\description{
Splits documents into text units
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chunk_size}}{Target chunk size}

\item{\code{chunk_overlap}}{Overlap size}

\item{\code{by_sentence}}{Preserve sentences}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DocumentChunker-new}{\code{DocumentChunker$new()}}
\item \href{#method-DocumentChunker-chunk}{\code{DocumentChunker$chunk()}}
\item \href{#method-DocumentChunker-clone}{\code{DocumentChunker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DocumentChunker-new"></a>}}
\if{latex}{\out{\hypertarget{method-DocumentChunker-new}{}}}
\subsection{Method \code{new()}}{
Create a new DocumentChunker
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentChunker$new(chunk_size = 1200, chunk_overlap = 100, by_sentence = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chunk_size}}{Target size}

\item{\code{chunk_overlap}}{Overlap}

\item{\code{by_sentence}}{Preserve sentences}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DocumentChunker-chunk"></a>}}
\if{latex}{\out{\hypertarget{method-DocumentChunker-chunk}{}}}
\subsection{Method \code{chunk()}}{
Chunk a document
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentChunker$chunk(text, document_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{Document text}

\item{\code{document_id}}{Document ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of TextUnit objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DocumentChunker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DocumentChunker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentChunker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
