% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedders.R
\name{download_vectors}
\alias{download_vectors}
\title{Download pre-trained word vectors}
\usage{
download_vectors(model = "glove-50", dest_dir = NULL)
}
\arguments{
\item{model}{Model name: "glove-50", "glove-100", "glove-200", "glove-300"}

\item{dest_dir}{Destination directory}
}
\value{
Path to downloaded model
}
\description{
Download GloVe or other pre-trained word vectors
}
