% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\docType{data}
\name{ENGLISH_STOPWORDS}
\alias{ENGLISH_STOPWORDS}
\title{English Stopwords}
\format{
An object of class \code{character} of length 45.
}
\usage{
ENGLISH_STOPWORDS
}
\description{
Common English stopwords
}
\keyword{datasets}
