% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{EnhancedSearchResults}
\alias{EnhancedSearchResults}
\title{Enhanced Search Results}
\description{
Search results with enterprise features
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{List of result items}

\item{\code{facets}}{Named list of FacetResult objects}

\item{\code{total_count}}{Total results before filtering}

\item{\code{filtered_count}}{Results after ACL filtering}

\item{\code{query_time_ms}}{Query time in milliseconds}

\item{\code{rerank_time_ms}}{Rerank time in milliseconds}

\item{\code{facet_time_ms}}{Facet time in milliseconds}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EnhancedSearchResults-new}{\code{EnhancedSearchResults$new()}}
\item \href{#method-EnhancedSearchResults-to_list}{\code{EnhancedSearchResults$to_list()}}
\item \href{#method-EnhancedSearchResults-clone}{\code{EnhancedSearchResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EnhancedSearchResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-EnhancedSearchResults-new}{}}}
\subsection{Method \code{new()}}{
Create new EnhancedSearchResults
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnhancedSearchResults$new(
  results,
  facets = list(),
  total_count = 0,
  filtered_count = 0,
  query_time_ms = 0,
  rerank_time_ms = 0,
  facet_time_ms = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{List of results}

\item{\code{facets}}{Named list of FacetResult}

\item{\code{total_count}}{Total count}

\item{\code{filtered_count}}{Filtered count}

\item{\code{query_time_ms}}{Query time}

\item{\code{rerank_time_ms}}{Rerank time}

\item{\code{facet_time_ms}}{Facet time}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EnhancedSearchResults-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EnhancedSearchResults-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnhancedSearchResults$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List representation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EnhancedSearchResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EnhancedSearchResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnhancedSearchResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
