% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Compute euclidean distance}
\usage{
euclidean_distance(query_vector, doc_vectors)
}
\arguments{
\item{query_vector}{Numeric vector (query)}

\item{doc_vectors}{Matrix of document vectors}
}
\value{
Numeric vector of distances
}
\description{
Compute euclidean distance
}
\keyword{internal}
