% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\name{ExtractionResult}
\alias{ExtractionResult}
\title{Extraction Result}
\description{
Result of entity extraction
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entities}}{List of Entity objects}

\item{\code{relationships}}{List of Relationship objects}

\item{\code{source_chunk}}{Source chunk ID}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExtractionResult-new}{\code{ExtractionResult$new()}}
\item \href{#method-ExtractionResult-clone}{\code{ExtractionResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExtractionResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExtractionResult-new}{}}}
\subsection{Method \code{new()}}{
Create new ExtractionResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExtractionResult$new(
  entities = list(),
  relationships = list(),
  source_chunk = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entities}}{Entities}

\item{\code{relationships}}{Relationships}

\item{\code{source_chunk}}{Source}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExtractionResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExtractionResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExtractionResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
