% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\docType{data}
\name{ExtractorType}
\alias{ExtractorType}
\title{Extractor Types}
\format{
An object of class \code{list} of length 4.
}
\usage{
ExtractorType
}
\description{
Extractor Types
}
\keyword{datasets}
