% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{FacetResult}
\alias{FacetResult}
\title{Facet Result}
\description{
Result of facet aggregation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field}}{Field name}

\item{\code{values}}{List of FacetValue objects}

\item{\code{total_count}}{Total count}

\item{\code{other_count}}{Count of values not in top-N}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FacetResult-new}{\code{FacetResult$new()}}
\item \href{#method-FacetResult-to_list}{\code{FacetResult$to_list()}}
\item \href{#method-FacetResult-clone}{\code{FacetResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-FacetResult-new}{}}}
\subsection{Method \code{new()}}{
Create a new FacetResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetResult$new(field, values, total_count, other_count = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Field name}

\item{\code{values}}{List of FacetValue objects}

\item{\code{total_count}}{Total count}

\item{\code{other_count}}{Other count}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetResult-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FacetResult-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetResult$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FacetResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
