% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{FacetValue}
\alias{FacetValue}
\title{Facet Value}
\description{
A single facet value with count
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{The facet value}

\item{\code{count}}{Number of occurrences}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FacetValue-new}{\code{FacetValue$new()}}
\item \href{#method-FacetValue-clone}{\code{FacetValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-FacetValue-new}{}}}
\subsection{Method \code{new()}}{
Create a new FacetValue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetValue$new(value, count)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value}

\item{\code{count}}{The count}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FacetValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
