% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_time}
\alias{format_time}
\title{Format time duration}
\usage{
format_time(ms)
}
\arguments{
\item{ms}{Time in milliseconds}
}
\value{
Formatted string
}
\description{
Format time duration
}
\keyword{internal}
