% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_path}
\alias{get_data_path}
\title{Get or create VectrixDB data directory}
\usage{
get_data_path(path = NULL)
}
\arguments{
\item{path}{Optional custom path}
}
\value{
Path to data directory
}
\description{
Get or create VectrixDB data directory
}
\keyword{internal}
