% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\name{GlobalSearcher}
\alias{GlobalSearcher}
\title{Global Searcher}
\description{
Community-based graph search
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{communities}}{List of communities}

\item{\code{k}}{Number of communities}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobalSearcher-new}{\code{GlobalSearcher$new()}}
\item \href{#method-GlobalSearcher-search}{\code{GlobalSearcher$search()}}
\item \href{#method-GlobalSearcher-clone}{\code{GlobalSearcher$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSearcher-new"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSearcher-new}{}}}
\subsection{Method \code{new()}}{
Create a new GlobalSearcher
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSearcher$new(communities, k = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{communities}}{List of Community objects}

\item{\code{k}}{Number of communities}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSearcher-search"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSearcher-search}{}}}
\subsection{Method \code{search()}}{
Search communities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSearcher$search(query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{Query string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GlobalSearchResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSearcher-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSearcher-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSearcher$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
