% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\name{GlobalSearchResult}
\alias{GlobalSearchResult}
\title{Global Search Result}
\description{
Result from global community search
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{communities}}{Matching communities}

\item{\code{summaries}}{Community summaries}

\item{\code{context}}{Combined context}

\item{\code{score}}{Relevance score}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobalSearchResult-new}{\code{GlobalSearchResult$new()}}
\item \href{#method-GlobalSearchResult-clone}{\code{GlobalSearchResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSearchResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSearchResult-new}{}}}
\subsection{Method \code{new()}}{
Create new GlobalSearchResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSearchResult$new(
  communities = list(),
  summaries = character(0),
  context = NULL,
  score = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{communities}}{Communities}

\item{\code{summaries}}{Summaries}

\item{\code{context}}{Context}

\item{\code{score}}{Score}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSearchResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSearchResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSearchResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
