% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\docType{data}
\name{GraphSearchType}
\alias{GraphSearchType}
\title{Graph Search Types}
\format{
An object of class \code{list} of length 3.
}
\usage{
GraphSearchType
}
\description{
Graph Search Types
}
\keyword{datasets}
