% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{KeywordAnalyzer}
\alias{KeywordAnalyzer}
\title{Keyword Analyzer}
\description{
Treats entire input as single token
}
\section{Super class}{
\code{\link[VectrixDB:TextAnalyzer]{VectrixDB::TextAnalyzer}} -> \code{KeywordAnalyzer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KeywordAnalyzer-analyze}{\code{KeywordAnalyzer$analyze()}}
\item \href{#method-KeywordAnalyzer-clone}{\code{KeywordAnalyzer$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="VectrixDB" data-topic="TextAnalyzer" data-id="analyze_query"><a href='../../VectrixDB/html/TextAnalyzer.html#method-TextAnalyzer-analyze_query'><code>VectrixDB::TextAnalyzer$analyze_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VectrixDB" data-topic="TextAnalyzer" data-id="initialize"><a href='../../VectrixDB/html/TextAnalyzer.html#method-TextAnalyzer-initialize'><code>VectrixDB::TextAnalyzer$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeywordAnalyzer-analyze"></a>}}
\if{latex}{\out{\hypertarget{method-KeywordAnalyzer-analyze}{}}}
\subsection{Method \code{analyze()}}{
Analyze text as single keyword
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeywordAnalyzer$analyze(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{Input text}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Single-element character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeywordAnalyzer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KeywordAnalyzer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeywordAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
