% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\docType{data}
\name{LLMProvider}
\alias{LLMProvider}
\title{LLM Provider Types}
\format{
An object of class \code{list} of length 4.
}
\usage{
LLMProvider
}
\description{
LLM Provider Types
}
\keyword{datasets}
