% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_vectors.R
\name{load_word2vec_binary}
\alias{load_word2vec_binary}
\title{Load word2vec binary format}
\usage{
load_word2vec_binary(path, max_words = NULL, normalize = TRUE)
}
\description{
Load word2vec binary format
}
\keyword{internal}
