% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_vectors.R
\name{load_word_vectors}
\alias{load_word_vectors}
\title{Load word vectors into memory}
\usage{
load_word_vectors(path, max_words = NULL, normalize = TRUE)
}
\arguments{
\item{path}{Path to word vectors file (GloVe .txt or word2vec .bin)}

\item{max_words}{Maximum number of words to load (NULL for all)}

\item{normalize}{Normalize vectors to unit length}
}
\value{
WordVectors object
}
\description{
Loads pre-trained word vectors from a file
}
