% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\name{LocalSearcher}
\alias{LocalSearcher}
\title{Local Searcher}
\description{
Entity-based graph search
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{graph}}{Knowledge graph}

\item{\code{k}}{Number of seed entities}

\item{\code{traversal_depth}}{Max hops}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LocalSearcher-new}{\code{LocalSearcher$new()}}
\item \href{#method-LocalSearcher-search}{\code{LocalSearcher$search()}}
\item \href{#method-LocalSearcher-clone}{\code{LocalSearcher$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSearcher-new"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSearcher-new}{}}}
\subsection{Method \code{new()}}{
Create a new LocalSearcher
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSearcher$new(graph, k = 10, traversal_depth = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{graph}}{KnowledgeGraph}

\item{\code{k}}{Seed entities}

\item{\code{traversal_depth}}{Max depth}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSearcher-search"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSearcher-search}{}}}
\subsection{Method \code{search()}}{
Search the graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSearcher$search(query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{Query string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
LocalSearchResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSearcher-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSearcher-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSearcher$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
