% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\name{LocalSearchResult}
\alias{LocalSearchResult}
\title{Local Search Result}
\description{
Result from local graph search
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entities}}{Matching entities}

\item{\code{relationships}}{Related relationships}

\item{\code{subgraph}}{Traversed subgraph}

\item{\code{context}}{Combined context text}

\item{\code{score}}{Relevance score}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LocalSearchResult-new}{\code{LocalSearchResult$new()}}
\item \href{#method-LocalSearchResult-clone}{\code{LocalSearchResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSearchResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSearchResult-new}{}}}
\subsection{Method \code{new()}}{
Create new LocalSearchResult
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSearchResult$new(
  entities = list(),
  relationships = list(),
  subgraph = NULL,
  context = NULL,
  score = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entities}}{Entities}

\item{\code{relationships}}{Relationships}

\item{\code{subgraph}}{SubGraph}

\item{\code{context}}{Context}

\item{\code{score}}{Score}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSearchResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSearchResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSearchResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
