% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mmr_rerank}
\alias{mmr_rerank}
\title{Maximal Marginal Relevance (MMR) reranking}
\usage{
mmr_rerank(
  query_vector,
  doc_vectors,
  doc_ids,
  scores,
  lambda = 0.7,
  limit = 10
)
}
\arguments{
\item{query_vector}{Query embedding}

\item{doc_vectors}{Matrix of document embeddings}

\item{doc_ids}{Vector of document IDs}

\item{scores}{Initial relevance scores}

\item{lambda}{Diversity parameter (0-1)}

\item{limit}{Number of results to return}
}
\value{
Data frame with reranked results
}
\description{
Maximal Marginal Relevance (MMR) reranking
}
\keyword{internal}
