% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_vectors}
\alias{normalize_vectors}
\title{Normalize vectors for cosine similarity}
\usage{
normalize_vectors(vectors)
}
\arguments{
\item{vectors}{Matrix of vectors (rows are vectors)}
}
\value{
Normalized matrix
}
\description{
Normalize vectors for cosine similarity
}
\keyword{internal}
