% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{parse_acl}
\alias{parse_acl}
\title{Parse ACL String}
\usage{
parse_acl(acl_string)
}
\arguments{
\item{acl_string}{ACL string}
}
\value{
ACLPrincipal object
}
\description{
Parse ACL string like 'user:alice' or 'group:engineering'
}
