% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{SearchMode}
\alias{SearchMode}
\title{Search Mode Enumeration}
\format{
An object of class \code{list} of length 5.
}
\usage{
SearchMode
}
\description{
Available search modes for VectrixDB
}
\keyword{datasets}
