% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{SimpleStemmer}
\alias{SimpleStemmer}
\title{Simple Stemmer}
\description{
Simple suffix-stripping stemmer (no external dependencies)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{suffixes}}{List of suffixes to remove}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimpleStemmer-stem}{\code{SimpleStemmer$stem()}}
\item \href{#method-SimpleStemmer-stem_words}{\code{SimpleStemmer$stem_words()}}
\item \href{#method-SimpleStemmer-clone}{\code{SimpleStemmer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimpleStemmer-stem"></a>}}
\if{latex}{\out{\hypertarget{method-SimpleStemmer-stem}{}}}
\subsection{Method \code{stem()}}{
Stem a word
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimpleStemmer$stem(word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{Word to stem}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Stemmed word
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimpleStemmer-stem_words"></a>}}
\if{latex}{\out{\hypertarget{method-SimpleStemmer-stem_words}{}}}
\subsection{Method \code{stem_words()}}{
Stem multiple words
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimpleStemmer$stem_words(words)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{words}}{Character vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Stemmed words
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimpleStemmer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimpleStemmer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimpleStemmer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
