% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphrag.R
\name{SubGraph}
\alias{SubGraph}
\title{SubGraph}
\description{
A subset of a knowledge graph
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entities}}{Entities in subgraph}

\item{\code{relationships}}{Relationships in subgraph}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SubGraph-new}{\code{SubGraph$new()}}
\item \href{#method-SubGraph-to_list}{\code{SubGraph$to_list()}}
\item \href{#method-SubGraph-clone}{\code{SubGraph$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubGraph-new"></a>}}
\if{latex}{\out{\hypertarget{method-SubGraph-new}{}}}
\subsection{Method \code{new()}}{
Create a new SubGraph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubGraph$new(entities = list(), relationships = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entities}}{Entities}

\item{\code{relationships}}{Relationships}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubGraph-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SubGraph-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubGraph$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubGraph-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SubGraph-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubGraph$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
