% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{text_analyzer_english}
\alias{text_analyzer_english}
\title{Create English Text Analyzer}
\usage{
text_analyzer_english()
}
\value{
TextAnalyzer object
}
\description{
English analyzer with stemming and stopwords
}
