% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{text_analyzer_keyword}
\alias{text_analyzer_keyword}
\title{Create Keyword Text Analyzer}
\usage{
text_analyzer_keyword()
}
\value{
TextAnalyzer object
}
\description{
No tokenization - treat input as single token
}
