% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{text_analyzer_simple}
\alias{text_analyzer_simple}
\title{Create Simple Text Analyzer}
\usage{
text_analyzer_simple()
}
\value{
TextAnalyzer object
}
\description{
Lowercase + letter-only tokenization
}
