% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{text_analyzer_standard}
\alias{text_analyzer_standard}
\title{Create Standard Text Analyzer}
\usage{
text_analyzer_standard()
}
\value{
TextAnalyzer object
}
\description{
Lowercase + basic tokenization
}
