% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_add_dir}
\alias{vdb_add_dir}
\title{Batch Add from Directory}
\usage{
vdb_add_dir(db, dir_path, pattern = "\\\\.txt$", recursive = TRUE)
}
\arguments{
\item{db}{Vectrix object or collection name}

\item{dir_path}{Directory path}

\item{pattern}{File pattern (default: "*.txt")}

\item{recursive}{Search subdirectories}
}
\value{
Vectrix object
}
\description{
Add all text files from a directory
}
\examples{
\dontrun{
vdb_add_dir(db, "./documents/")
}
}
