% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_create}
\alias{vdb_create}
\title{Create Collection}
\usage{
vdb_create(name, model = "tfidf", dimension = NULL, data_dir = NULL)
}
\arguments{
\item{name}{Collection name}

\item{model}{Embedding model}

\item{dimension}{Vector dimension}

\item{data_dir}{Data directory}
}
\value{
Vectrix object
}
\description{
Create a new VectrixDB collection
}
\examples{
\dontrun{
db <- vdb_create("my_docs")
}
}
