% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz_dashboard_exact.R
\name{vdb_dashboard_simple}
\alias{vdb_dashboard_simple}
\title{Launch Simple Dashboard}
\usage{
vdb_dashboard_simple(db)
}
\arguments{
\item{db}{Vectrix object.}
}
\value{
Invisibly returns server object from \code{vdb_dashboard()}.
}
\description{
Convenience wrapper around \code{vdb_dashboard()} using \code{db$path}.
}
