% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_delete_docs}
\alias{vdb_delete_docs}
\title{Delete Documents}
\usage{
vdb_delete_docs(db, ids)
}
\arguments{
\item{db}{Vectrix object or collection name}

\item{ids}{Document ID(s)}
}
\value{
Vectrix object
}
\description{
Delete documents by ID
}
