% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_get}
\alias{vdb_get}
\title{Get Document}
\usage{
vdb_get(db, ids)
}
\arguments{
\item{db}{Vectrix object or collection name}

\item{ids}{Document ID(s)}
}
\value{
List of Result objects
}
\description{
Get document by ID
}
