% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_import}
\alias{vdb_import}
\title{Import from File}
\usage{
vdb_import(db, path, separator = "\\n")
}
\arguments{
\item{db}{Vectrix object or collection name}

\item{path}{Input file path}

\item{separator}{Line separator for documents}
}
\value{
Vectrix object
}
\description{
Import documents from text file
}
\examples{
\dontrun{
vdb_import(db, "documents.txt")
}
}
