% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_info}
\alias{vdb_info}
\title{Collection Info}
\usage{
vdb_info(db)
}
\arguments{
\item{db}{Vectrix object or collection name}
}
\value{
Named list of info
}
\description{
Get collection information
}
\examples{
\dontrun{
vdb_info(db)
vdb_info("my_docs")
}
}
