% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_interactive}
\alias{vdb_interactive}
\title{Start Interactive CLI}
\usage{
vdb_interactive(collection = NULL)
}
\arguments{
\item{collection}{Default collection name}
}
\description{
Start an interactive VectrixDB session
}
\examples{
\dontrun{
vdb_interactive()
}
}
