% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_list}
\alias{vdb_list}
\title{List Collections}
\usage{
vdb_list(data_dir = NULL)
}
\arguments{
\item{data_dir}{Data directory path}
}
\value{
Character vector of collection names
}
\description{
List all VectrixDB collections in the data directory
}
\examples{
\dontrun{
vdb_list()
}
}
