% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_open}
\alias{vdb_open}
\title{Open Collection}
\usage{
vdb_open(name, data_dir = NULL)
}
\arguments{
\item{name}{Collection name}

\item{data_dir}{Data directory}
}
\value{
Vectrix object
}
\description{
Open an existing collection
}
\examples{
\dontrun{
db <- vdb_open("my_docs")
}
}
