% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vdb_stats}
\alias{vdb_stats}
\title{Collection Statistics}
\usage{
vdb_stats(db)
}
\arguments{
\item{db}{Vectrix object or collection name}
}
\value{
Named list of stats
}
\description{
Get detailed statistics
}
