% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectrix.R
\name{vectrix_create}
\alias{vectrix_create}
\title{Create a new Vectrix collection}
\usage{
vectrix_create(name = "default", ...)
}
\arguments{
\item{name}{Collection name}

\item{...}{Additional arguments passed to Vectrix$new()}
}
\value{
Vectrix object
}
\description{
Create a new Vectrix collection
}
