% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectrix.R
\name{vectrix_open}
\alias{vectrix_open}
\title{Open an existing Vectrix collection}
\usage{
vectrix_open(name = "default", path = NULL)
}
\arguments{
\item{name}{Collection name}

\item{path}{Storage path}
}
\value{
Vectrix object
}
\description{
Open an existing Vectrix collection
}
