% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{word2vec_available}
\alias{word2vec_available}
\title{Check whether word2vec package is installed}
\usage{
word2vec_available()
}
\value{
Logical
}
\description{
Check whether word2vec package is installed
}
\keyword{internal}
