% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{DistanceMetric}
\alias{DistanceMetric}
\title{Distance Metric Enumeration}
\format{
An object of class \code{list} of length 4.
}
\usage{
DistanceMetric
}
\description{
Available distance metrics for vector comparison
}
\keyword{datasets}
