% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{vectrix_info}
\alias{vectrix_info}
\title{Display VectrixDB information}
\usage{
vectrix_info(path = "./vectrixdb_data")
}
\arguments{
\item{path}{Database path}
}
\description{
Show database statistics and info
}
\examples{
\dontrun{
vectrix_info("./my_data")
}
}
